refactor(audio-capture): consolidate vendored crates
