+++ packages/audio-capture/crates/audio-actual/Cargo.toml
@@ -10,8 +10,6 @@ license = "MIT"
 
 [dependencies]
 hypr-aec = { workspace = true }
-hypr-audio = { workspace = true }
-hypr-audio-interface = { workspace = true }
 hypr-audio-sync = { workspace = true }
 hypr-audio-utils = { workspace = true }
 hypr-resampler = { workspace = true }
@@ -21,6 +19,7 @@ tracing = { workspace = true }
 
 futures-util = { workspace = true }
 pin-project = { workspace = true }
+thiserror = { workspace = true }
 tokio = { workspace = true, features = ["rt", "macros", "sync", "time"] }
 tokio-stream = { workspace = true }
 tokio-util = { workspace = true }
