+++ packages/audio-capture/crates/audio-actual/src/capture/mod.rs
@@ -4,7 +4,7 @@
 mod joiner;
 mod stream;
 
-use hypr_audio::{CaptureConfig, CaptureStream, Error};
+use crate::{CaptureConfig, CaptureStream, Error};
 use stream::{CaptureSide, setup_mic_stream, setup_speaker_stream};
 
 pub(crate) fn open_capture(config: CaptureConfig) -> Result<CaptureStream, Error> {
