+++ packages/audio-capture/crates/audio-actual/src/speaker/macos.rs
@@ -7,8 +7,8 @@ use std::task::Poll;
 use anyhow::Result;
 use futures_util::Stream;
 use futures_util::task::AtomicWaker;
-use hypr_audio_interface::AsyncSource;
 use hypr_audio_utils::{pcm_f64_to_f32, pcm_i16_to_f32, pcm_i32_to_f32};
+use hypr_resampler::AsyncSource;
 use pin_project::pin_project;
 
 use crate::async_ring::RingbufAsyncReader;
