--- packages/audio-capture/crates/audio-interface/Cargo.toml
@@ -1,10 +1,0 @@
-# Vendored from https://github.com/fastrepl/hyprnote (crates/audio-interface), MIT licensed.
-# See crates/LICENSE for the upstream license. The optional rodio integration was dropped.
-[package]
-name = "hypr-audio-interface"
-version = "0.1.0"
-edition = "2024"
-license = "MIT"
-
-[dependencies]
-futures-util = { workspace = true }
