--- packages/audio-capture/crates/audio/Cargo.toml
@@ -1,11 +1,0 @@
-# Vendored from https://github.com/fastrepl/hyprnote (crates/audio), MIT licensed.
-# See crates/LICENSE for the upstream license.
-[package]
-name = "hypr-audio"
-version = "0.1.0"
-edition = "2024"
-license = "MIT"
-
-[dependencies]
-futures-util = { workspace = true }
-thiserror = { workspace = true }
