+++ packages/audio-capture/crates/resampler/src/async_source/dynamic_new.rs
@@ -4,10 +4,9 @@ use std::pin::Pin;
 use std::task::{Context, Poll};
 
 use futures_util::{Stream, pin_mut};
-use hypr_audio_interface::AsyncSource;
 use pin_project::pin_project;
 
-use crate::{Async, FixedAsync, PolynomialDegree, RubatoChunkResampler};
+use crate::{Async, AsyncSource, FixedAsync, PolynomialDegree, RubatoChunkResampler};
 
 pub trait ResampleExtDynamicNew: AsyncSource + Sized {
   fn resampled_chunks(
