+++ packages/audio-capture/crates/vad-masking/src/lib.rs
@@ -1,5 +1,6 @@
 // Vendored from https://github.com/fastrepl/hyprnote (crates/vad-masking/src/lib.rs), MIT licensed.
 
+mod earshot;
 mod masking;
 mod streaming;
