+++ packages/audio-capture/crates/vad-masking/src/streaming.rs
@@ -1,7 +1,8 @@
 // Vendored from https://github.com/fastrepl/hyprnote (crates/vad-masking/src/streaming.rs), MIT licensed.
 
 use hypr_audio_utils::f32_to_i16_samples;
-use hypr_vad::earshot::VoiceActivityDetector;
+
+use crate::earshot::{VoiceActivityDetector, choose_optimal_frame_size};
 
 #[derive(Clone, Debug)]
 pub struct VadConfig {
@@ -35,7 +36,7 @@ impl StreamingVad {
   }
 
   pub fn with_config(frame_hint: usize, cfg: VadConfig) -> Self {
-    let frame_size = hypr_vad::earshot::choose_optimal_frame_size(frame_hint);
+    let frame_size = choose_optimal_frame_size(frame_hint);
     debug_assert!(frame_size > 0, "VAD frame size must be > 0");
 
     Self {
