--- packages/audio-capture/crates/vad/src/lib.rs
@@ -1,4 +1,0 @@
-// Vendored from https://github.com/fastrepl/hyprnote (crates/vad/src/lib.rs), MIT licensed.
-// Trimmed to the earshot backend; the silero-onnx backend was dropped.
-
-pub mod earshot;
